Added GPS Coordinates for stations table
And updated all necessary procedures

+++ src/benchmarks/edu/brown/benchmark/bikerstream/bikerstream-ddl.sql
@@ -8,8 +8,8 @@ CREATE TABLE stations
     station_id     INTEGER      PRIMARY KEY
 ,   station_name   VARCHAR(64)  NOT NULL
 ,   street_address VARCHAR(128) NOT NULL
-,   latitude       INTEGER      NOT NULL
-,   longitude      INTEGER      NOT NULL
+,   latitude       FLOAT        NOT NULL
+,   longitude      FLOAT        NOT NULL
 );
 
 CREATE TABLE StationStatus
